feat: ToSettle Logic && fix ConsultingHistory Logic

+++ VoiceConsultingApp/VoiceConsultingApp/Model/SettlementDetail/SettlementDetail.swift
@@ -13,6 +13,6 @@ struct SettlementDetail: Codable {
     var coinCount: Int
     var price: Int
     var isSettle: Bool
-    var createAt: Double
+    var createAt: Int
     var accountInfo: String
 }
